MyBookNotes/requirements.txt
aiogram==2.25.1
SQLAlchemy==2.0.23
pydantic-settings==0.0.1
python-dotenv==1.0.0
